Refactor separate log_output_utils.go
